# dealer-server/src/main/java/com/autodealer/crm/dto/CustomerDuplicateResponse.java
package com.autodealer.crm.dto;

import java.util.ArrayList;
import java.util.List;
import lombok.Data;

@Data
public class CustomerDuplicateResponse {

    private boolean duplicated;

    private boolean hiddenConflict;

    private List<CustomerDuplicateSummary> visibleCustomers = new ArrayList<>();
}