# dealer-server/src/main/java/com/autodealer/crm/dto/CustomerDuplicateSummary.java
package com.autodealer.crm.dto;

import lombok.Data;

@Data
public class CustomerDuplicateSummary {

    private Integer customerId;

    private String customerName;

    private String maskedPhone;

    private String ownerName;
}