# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: 汽车经销商 CRM API
  version: 1.0.0
  description: 当前后端 Controller 暴露接口契约。统一规则见 docs/api/api-conventions.md。
servers:
  - url: http://localhost:8089
    description: 本地后端服务
security:
  - BearerAuth: []
...
paths:
...
  /api/customer/list:
    get:
      tags: [Customer]
      operationId: listCustomers
      summary: 分页查询客户列表
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
        - name: customerName
          in: query
          schema:
            type: string
        - name: productId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          $ref: "#/components/responses/CustomerPageOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/customer/options:
    get:
      tags: [Customer]
      operationId: listCustomerOptions
      summary: 查询客户下拉选项
      responses:
        "200":
          $ref: "#/components/responses/CustomerOptionArrayOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/customer/{id}:
    get:
      tags: [Customer]
      operationId: getCustomer
      summary: 查询客户详情
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/CustomerDetailObject"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [Customer]
      operationId: deleteCustomer
      summary: 删除无业务关系客户
      description: 删除前检查线索、跟进、报价、交易等业务关系；存在任一引用时返回 422 RESOURCE_IN_USE，不物理删除。
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/customer/{id}/owner:
    put:
      tags: [Customer]
      operationId: transferCustomerOwner
      summary: 转移客户负责人
      description: 按当前用户数据范围校验客户，目标负责人必须有效，转移原因必填，并写入客户归属历史。
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferCustomerOwnerRequest"
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/customer/{id}/merge:
    post:
      tags: [Customer]
      operationId: mergeCustomer
      summary: 合并重复客户
      description: 将被合并客户的跟进、交易和报价迁移到主客户，被合并客户标记为 MERGED 并保留合并事实。
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeCustomerRequest"
      responses:
        "200":
          $ref: "#/components/responses/CustomerMergeObject"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/clue/customer:
    post:
      tags: [Customer]
      operationId: convertClueToCustomer
      summary: 线索转客户
      description: 将有权访问的线索转为客户主档事实；不自动创建交易、报价、订单、收款、发票或库存占用。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertCustomerRequest"
      responses:
        "200":
          $ref: "#/components/responses/EmptyOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/CustomerDuplicateConflict"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/customers:
    get:
      tags: [Customer]
      operationId: listCustomersDeprecated
      summary: 分页查询客户列表，历史接口
      deprecated: true
      parameters:
        - $ref: "#/components/parameters/Current"
      responses:
        "200":
          $ref: "#/components/responses/PageOk"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/exportExcel:
    get:
      tags: [Customer]
      operationId: exportCustomers
      summary: 导出客户 Excel
      parameters:
        - name: ids
          in: query
          required: false
          schema:
            type: string
          description: 逗号分隔客户 ID；为空时导出可见范围内客户。
      responses:
        "200":
          description: Excel 文件。
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    Current:
      name: current
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
      description: 历史分页页码。
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 1
        minimum: 1
      description: 分页页码。
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 10
        minimum: 1
      description: 分页大小。
...
  responses:
    EmptyOk:
      description: 操作成功，无业务数据。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/R"
...
    PageOk:
      description: 分页查询成功，data 为 PageInfo。
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/R"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PageInfo"
    CustomerPageOk:
      description: 分页查询成功，data.list 为客户列表响应；敏感字段由后端按权限脱敏。
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/R"
              - type: object
                properties:
                  data:
                    allOf:
                      - $ref: "#/components/schemas/PageInfo"
                      - type: object
                        properties:
                          list:
                            type: array
                            items:
                              $ref: "#/components/schemas/CustomerListResponse"
    CustomerOptionArrayOk:
      description: 查询成功，data 为客户下拉选项数组。
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/R"
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/CustomerOption"
    CustomerDetailObject:
      description: 查询成功，data 为客户详情；敏感字段由后端按权限脱敏。
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/R"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CustomerDetailResponse"
    CustomerMergeObject:
      description: 合并成功，data 为本次迁移数量。
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/R"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CustomerMergeResponse"
    CustomerDuplicateConflict:
      description: 客户联系方式疑似重复；越权重复只返回 hiddenConflict，不泄露客户敏感信息。
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorR"
              - type: object
                properties:
                  data:
                    $ref: "#/components/schemas/CustomerDuplicateResponse"
...
    BadRequest:
      description: 请求参数错误。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    Unauthorized:
      description: 未认证、Token 无效或 Token 过期。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    Forbidden:
      description: 功能权限或数据权限不足。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    NotFound:
      description: 资源不存在。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    Conflict:
      description: 重复、并发冲突或非法状态迁移。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    Unprocessable:
      description: 请求格式正确但业务对象无法处理。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
    ServerError:
      description: 未预期系统异常。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorR"
  schemas:
    R:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: 操作成功
        data:
          nullable: true
    ErrorR:
      type: object
      required: [code, msg]
      properties:
        code:
          type: integer
          example: 501
        msg:
          type: string
          example: 请求参数格式有误
        data:
          nullable: true
    PageInfo:
      type: object
      properties:
        pageNum:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
          format: int64
        pages:
          type: integer
        list:
          type: array
          items:
            type: object
            additionalProperties: true
...
    ConvertCustomerRequest:
      type: object
      required: [clueId]
      properties:
        clueId:
          type: integer
        product:
          type: integer
          format: int64
        quantity:
          type: integer
          default: 1
        description:
          type: string
        nextContactTime:
          type: string
          format: date-time
    CustomerListResponse:
      type: object
      properties:
        id:
          type: integer
        clueId:
          type: integer
        customerName:
          type: string
        phone:
          type: string
          description: 无 customer:sensitive:view 权限时由后端脱敏。
        weixin:
          type: string
          description: 无 customer:sensitive:view 权限时由后端脱敏。
        ownerName:
          type: string
        activityName:
          type: string
        appellationName:
          type: string
        needLoanName:
          type: string
        intentionStateName:
          type: string
        stateName:
          type: string
          description: 兼容字段，等同 customerStatusName。
        sourceName:
          type: string
        originalSourceName:
          type: string
        customerStatus:
          type: string
          enum: [POTENTIAL, INTENTION, VISITED, TEST_DRIVE, DEAL, REPURCHASE, DORMANT, LOST, INVALID, MERGED]
        customerStatusName:
          type: string
        intentionProductName:
          type: string
        product:
          type: integer
          format: int64
        description:
          type: string
        nextContactTime:
          type: string
          format: date-time
        createTime:
          type: string
          format: date-time
    CustomerDetailResponse:
      allOf:
        - $ref: "#/components/schemas/CustomerListResponse"
        - type: object
          properties:
            qq:
              type: string
              description: 无 customer:sensitive:view 权限时由后端脱敏。
            email:
              type: string
              description: 无 customer:sensitive:view 权限时由后端脱敏。
            age:
              type: integer
            job:
              type: string
            yearIncome:
              type: string
            address:
              type: string
              description: 无 customer:sensitive:view 权限时由后端脱敏。
            productName:
              type: string
    CustomerOption:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        clueId:
          type: integer
    TransferCustomerOwnerRequest:
      type: object
      required: [newOwnerId, reason]
      properties:
        newOwnerId:
          type: integer
        reason:
          type: string
          maxLength: 255
    MergeCustomerRequest:
      type: object
      required: [sourceCustomerId, reason]
      properties:
        sourceCustomerId:
          type: integer
        reason:
          type: string
          maxLength: 255
    CustomerMergeResponse:
      type: object
      properties:
        targetCustomerId:
          type: integer
        sourceCustomerId:
          type: integer
        migratedRemarkCount:
          type: integer
        migratedTranCount:
          type: integer
        migratedQuoteCount:
          type: integer
    CustomerDuplicateSummary:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        maskedPhone:
          type: string
        ownerName:
          type: string
    CustomerDuplicateResponse:
      type: object
      properties:
        duplicated:
          type: boolean
        hiddenConflict:
          type: boolean
        visibleCustomers:
          type: array
          items:
            $ref: "#/components/schemas/CustomerDuplicateSummary"
...